Restructure OIDC production code
