+++ app/domain/authentication/authn_oidc/authenticate_id_token/authenticate_request_body.rb
@@ -1,0 +1,27 @@
+require 'uri'
+require 'json'
+
+module Authentication
+  module AuthnOidc
+    module AuthenticateIdToken
+      class AuthenticateRequestBody
+        attr_reader :id_token
+
+        def initialize(request)
+          @id_token = parsed_id_token(request.body.read)
+        end
+
+        private
+
+        def decoded_body(request_body)
+          URI.decode_www_form(request_body)
+        end
+
+        def parsed_id_token(request_body)
+          decoded_request_body = decoded_body(request_body)
+          decoded_request_body.assoc('id_token').last
+        end
+      end
+    end
+  end
+end
