--- app/domain/authentication/authn_oidc/authenticate_request_body.rb
@@ -1,25 +1,0 @@
-require 'uri'
-require 'json'
-
-module Authentication
-  module AuthnOidc
-    class AuthenticateRequestBody
-      attr_reader :id_token
-
-      def initialize(request)
-        @id_token = parsed_id_token(request.body.read)
-      end
-
-      private
-
-      def decoded_body(request_body)
-        URI.decode_www_form(request_body)
-      end
-
-      def parsed_id_token(request_body)
-        decoded_request_body = decoded_body(request_body)
-        decoded_request_body.assoc('id_token').last
-      end
-    end
-  end
-end
